requirements.txt
streamlit
plotly
pandas
numpy
dnspython
scapy
pyshark
loguru
tqdm
scikit-learn
imbalanced-learn
xgboost
matplotlib
colorama
pyyaml
joblib
requests
pytest
